\name{dim.neurocdf}
\alias{dim.neurocdf}
\title{Extract dimensions of neuro netCDF file}
\usage{
  \method{dim}{neurocdf} (x, ...)
}
\arguments{
  \item{x}{neuro netCDF filename}

  \item{...}{Additional arguments parsed to lower level
  functions}
}
\description{
  Extract dimensions of neuro netCDF file
}
\author{
  Klaus K. Holst
}

