\name{id.neurocdf}
\alias{id.neurocdf}
\title{Extract different 'id' of neuro netCDF files}
\usage{
  \method{id}{neurocdf} (object, ...)
}
\arguments{
  \item{object}{netCDF filename}

  \item{...}{Additional low level argument parsed on to
  lower level functions}
}
\description{
  Extract different 'id' of neuro netCDF files
}
\author{
  Klaus K. Holst
}

