\name{mkNeuro}
\alias{mkNeuro}
\title{Create brain volume or slice}
\usage{
  mkNeuro(object, value, plane = NULL, slice, ...)
}
\arguments{
  \item{object}{neuro netCDF filename or array of voxels}

  \item{value}{Image values (optional)}

  \item{plane}{Plane to alter (optional)}

  \item{slice}{Slice to alter (optional)}

  \item{...}{Additional arguments parsed on to lower level
  functions}
}
\description{
  Create brain volume or slice
}
\author{
  Klaus K. Holst
}

