\docType{package}
\name{neurocdf-package}
\alias{neurocdf-package}
\title{Various utilities for handling neuroimaging data}
\description{
  Various utilities for handling neuroimaging data. In
  particular methods for handling neuroimaging data via the
  netCDF format is available.
}
\examples{
## To appear
}
\author{
  Klaus K. Holst
}
\keyword{package}

