\name{summary.neurocdf}
\alias{summary.neurocdf}
\title{Summary of neuro netCDF file}
\usage{
  \method{summary}{neurocdf} (object, ...)
}
\arguments{
  \item{object}{neuro netCDF filename}

  \item{...}{Additional arguments parsed to lower level
  functions}
}
\description{
  Summary of neuro netCDF file
}
\author{
  Klaus K. Holst
}

