\name{type.neurocdf}
\alias{type}
\alias{type.neurocdf}
\title{Extract different 'types' of neuro netCDF files}
\usage{
  \method{type}{neurocdf} (object, ...)
}
\arguments{
  \item{object}{netCDF filename}

  \item{...}{Additional low level argument parsed on to
  lower level functions}
}
\description{
  Extract different 'types' of neuro netCDF files
}
\author{
  Klaus K. Holst
}

